transaction_verification/requirements.txt
grpcio==1.43.0
grpcio-tools==1.43.0
protobuf==3.19.4
watchdog==2.1.6